Update state data model to use pod spec instead of container spec
